+++ package.json
@@ -27,7 +27,6 @@
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-mobile-picker": "^1.2.0",
-    "react-modal-sheet": "^5.6.0",
     "sonner": "^2.0.7",
     "styled-components": "^6.1.19",
     "styled-system": "^5.1.5",
