+++ src/app/(protected)/dashboard/settings/page.tsx
@@ -1,0 +1,5 @@
+import SettingsScreen from "@/widgets/settingsScreen";
+
+export default function SettingsPage() {
+  return <SettingsScreen />;
+}
