--- src/features/settings/components/index.ts
@@ -1,1 +1,0 @@
-export { default as SettingsBottomSheet } from "./settingsBottomSheet";
