+++ src/shared/lib/navigation/dashboard.ts
@@ -1,4 +1,9 @@
 export const DASHBOARD_PATH = "/dashboard";
+export const DASHBOARD_SETTINGS_PATH = `${DASHBOARD_PATH}/settings`;
+
+export function isDashboardSettingsPath(pathname: string) {
+  return pathname === DASHBOARD_SETTINGS_PATH;
+}
 
 export function isDashboardFormPath(pathname: string) {
   return (
